+++ plugins/orion/src/bak/workflow/index.ts
@@ -1,3 +1,2 @@
-export { WorkFlowDetail } from './WorkFlowDetail';
 export { Workflow } from './Workflow';
 export { Onboarding } from './Onboarding';
