+++ plugins/orion/src/components/workflow/Workflow.tsx
@@ -1,203 +1,70 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   ContentHeader,
-  ItemCardHeader,
+  InfoCard,
+  Progress,
   SupportButton,
 } from '@backstage/core-components';
-import {
-  Button,
-  Card,
-  CardActions,
-  CardContent,
-  CardMedia,
-  Grid,
-  makeStyles,
-  OutlinedInputProps,
-  TextField,
-  Typography,
-} from '@material-ui/core';
-
-import { useCommonStyles } from '../../styles';
+import { useEffect } from 'react';
+import { errorApiRef, useApi } from '@backstage/core-plugin-api';
+import { Form } from '../Form/Form';
 import { ParodosPage } from '../ParodosPage';
+import { Typography, Button, makeStyles, Grid } from '@material-ui/core';
+import { useGetProjectAssessmentSchema } from './useGetProjectAssessmentSchema';
+import type { AssessmentStatusType } from '../types';
+import useAsyncFn from 'react-use/lib/useAsyncFn';
 import { useBackendUrl } from '../api';
-import {
-  AssessmentStatusType,
-  ProjectType,
-  WorkflowDefinitionType,
-  WorkFlowTaskParameterType,
-} from '../types';
-import { mockAndromedaWorkflowDefinition } from './mockData';
-import { WorkflowParameterComponent } from './WorkflowParameterComponent';
-import { getWorkflowParameters, startWorkflow } from './commands';
-import {
-  WorkflowParametersContext,
-  WorkflowParametersContextProvider,
-} from '../../context/WorkflowParametersContext';
+import { IChangeEvent } from '@rjsf/core-v5';
+import { type Project } from '../../models/workflowDefinitionSchema';
+import { WorkflowDefinitions } from './WorkflowDefinitions';
+import { useGetWorkflowDefinitions } from '../../hooks/useGetWorkflowDefinitions';
 
 const useStyles = makeStyles({
-  applicationHeader: {
-    background: 'gray',
-  },
-  applicationCard: {
-    height: '15rem',
-    width: '17rem',
+  fullHeight: {
+    height: '100%',
   },
 });
 
-const WorkflowImpl: React.FC = () => {
-  const commonStyles = useCommonStyles();
-  const styles = useStyles();
+export function Workflow(): JSX.Element {
+  const [project, setProject] = useState<Project>();
   const backendUrl = useBackendUrl();
-
-  const [projectName, setProjectName] = React.useState<string>('');
-  const { getParamValue, getParamValidation } = React.useContext(
-    WorkflowParametersContext,
-  );
-  const [project, setProject] = React.useState<ProjectType>();
+  const { value: workflowDefinitions = [] } = useGetWorkflowDefinitions();
   const [assessmentStatus, setAssessmentStatus] =
-    React.useState<AssessmentStatusType>('none');
-  const [assessmentWorkflowDefinition, setAssessmentWorkflowDefinition] =
-    React.useState<WorkflowDefinitionType>();
-  const [workflowDefinitions, setWorkflowDefinitions] = React.useState<
-    WorkflowDefinitionType[]
-  >([]);
-  const [assessmentParameters, setAssessmentParameters] = React.useState<
-    WorkFlowTaskParameterType[]
-  >([]);
-  const [_, setError] = React.useState<string>();
+    useState<AssessmentStatusType>('none');
+  const styles = useStyles();
 
-  const onChangeProjectName: OutlinedInputProps['onChange'] = event => {
-    setProjectName(event.target.value);
-  };
+  const { loading, error, value: formSchema } = useGetProjectAssessmentSchema();
 
-  // Get a list of the ASSESSMENT workflow parameters to ask the user for
-  React.useEffect(() => {
-    const doItAsync = async () => {
-      const response = await fetch(
-        `${backendUrl}/api/proxy/parodos/workflowdefinitions`,
-      );
-      const allWorkflowDefinitions =
-        (await response.json()) as WorkflowDefinitionType[];
-      const assessmentDefinition = allWorkflowDefinitions.find(
-        def => def.type === 'ASSESSMENT',
-      );
-
-      if (!assessmentDefinition) {
-        setError('Could not find assessment definition.');
-        return;
-      }
-
-      const params = getWorkflowParameters(
-        allWorkflowDefinitions,
-        assessmentDefinition,
-      );
-
-      setAssessmentWorkflowDefinition(assessmentDefinition);
-      setAssessmentParameters(params);
-    };
-    doItAsync();
-  }, [backendUrl]);
-
-  const onStartAssessment = () => {
-    // TODO: get the state dynmically
+  const [, startAssessment] = useAsyncFn(async ({ formData }: IChangeEvent) => {
     setAssessmentStatus('inprogress');
 
-    const doItAsync = async () => {
-      if (!assessmentWorkflowDefinition) {
-        setError('Missing assessment definition.');
-        return;
-      }
-
-      try {
-        // TODO: shouldn't following be executed as a part of the Assessment workflow?
-        // It was said to call following but should be part of the backend flow, imho
-        const response = await fetch(
-          `${backendUrl}/api/proxy/parodos/projects`,
-          {
-            method: 'POST',
-            body: JSON.stringify({
-              name: projectName,
-            }),
-          },
-        );
-        const prj = (await response.json()) as ProjectType;
-        setProject(prj);
-
-        // Start the assessment workflow
-        const executionId = await startWorkflow({
-          workflow: assessmentWorkflowDefinition,
-          projectId: prj.id,
-          getParamValue,
-          backendUrl,
-          setError,
-        });
-
-        // TODO: Implement Cancel - https://issues.redhat.com/browse/FLPATH-101
-        // eslint-disable-next-line no-console
-        console.log(
-          'Assessment executionId: ',
-          executionId,
-          ', do something about it.',
-        );
-
-        // TODO: replace following by proper monitoring
-        // https://issues.redhat.com/browse/FLPATH-100
-        setAssessmentStatus('complete');
-      } catch (e) {
-        setError('Failed to start assessment.');
-        // eslint-disable-next-line no-console
-        console.error('Error: ', e);
-      }
-    };
-    doItAsync();
-  };
-
-  // Read workflows to display once the Assessment is over
-  React.useEffect(() => {
-    const doItAsync = async () => {
-      try {
-        const response = await fetch(
-          `${backendUrl}/api/proxy/parodos/workflowdefinitions`,
-        );
-        const allWorkflowDefinitions =
-          (await response.json()) as WorkflowDefinitionType[];
-
-        let filteredWorkflowDefinitions = allWorkflowDefinitions.filter(
-          workflowDefinition =>
-            // TODO: is following correct?
-            !['ASSESSMENT', 'CHECKER'].includes(workflowDefinition.type),
-        );
-
-        // mock - TODO: remove
-        filteredWorkflowDefinitions = [
-          mockAndromedaWorkflowDefinition,
-          ...filteredWorkflowDefinitions,
-        ];
-
-        setWorkflowDefinitions(filteredWorkflowDefinitions);
-      } catch (e) {
-        setError('Failed to get workflow definitions.');
-        // eslint-disable-next-line no-console
-        console.error('Error: ', e);
-      }
-    };
-    doItAsync();
-  }, [backendUrl]);
-
-  const isStartDisabled =
-    assessmentStatus !== 'none' ||
-    !projectName ||
-    !!assessmentParameters.find(param => {
-      // Make sure all required fields are entered
-      if (!param.optional && !getParamValue(param.key)) {
-        return true;
-      }
-
-      return !!getParamValidation(param.key);
+    const response = await fetch(`${backendUrl}/api/proxy/parodos/projects`, {
+      method: 'POST',
+      body: JSON.stringify({
+        name: formData.projectName,
+      }),
     });
 
+    setProject((await response.json()) as Project);
+
+    setAssessmentStatus('complete');
+  }, []);
+
+  const errorApi = useApi(errorApiRef);
+
+  useEffect(() => {
+    if (error) {
+      errorApi.post(new Error(`Getting definition failed, ${error}`));
+    }
+  }, [error, errorApi]);
+
+  const inProgress = assessmentStatus === 'inprogress';
+  const complete = assessmentStatus === 'complete';
+
+  const disableForm = inProgress || complete;
+
   return (
-    <ParodosPage>
+    <ParodosPage stretch>
       <ContentHeader title="Project assessment">
         <SupportButton title="Need help?">Lorem Ipsum</SupportButton>
       </ContentHeader>
@@ -205,99 +72,43 @@ const WorkflowImpl: React.FC = () => {
         Select a project for an assessment of what additional workflows, if any,
         it qualifies for.
       </Typography>
-
-      <Grid container direction="column" spacing={2}>
-        <Grid container direction="row" spacing={2}>
-          <Grid item>
-            {/* This one seems to be special, on top of the Assessment workflow */}
-            <TextField
-              id="project-name"
-              disabled={assessmentStatus !== 'none'}
-              label="Project name"
-              variant="outlined"
-              value={projectName}
-              onChange={onChangeProjectName}
-            />
-          </Grid>
-
-          <Grid container spacing={3}>
-            {assessmentParameters.map(param => (
-              <Grid item xs={2} key={param.key}>
-                <WorkflowParameterComponent param={param} />
-              </Grid>
-            ))}
-          </Grid>
-
-          <Grid item className={commonStyles.paddingtop1}>
-            {assessmentStatus === 'inprogress' ? (
-              <Button
-                id="assessment-inprogress"
-                disabled
-                variant="contained"
-                color="primary"
+      {loading && <Progress />}
+      {formSchema?.schema && (
+        <InfoCard className={styles.fullHeight}>
+          <Grid container direction="row">
+            <Grid item xs={12} xl={8}>
+              <Form
+                formSchema={formSchema}
+                onSubmit={startAssessment}
+                disabled={disableForm}
               >
-                IN PROGRESS
-              </Button>
-            ) : (
-              <Button
-                id="assessment-start"
-                disabled={isStartDisabled}
-                variant="contained"
-                onClick={onStartAssessment}
-                color="primary"
-              >
-                START ASSESSMENT
-              </Button>
-            )}
-          </Grid>
-
-          <Grid item>{/* Space saver */}</Grid>
-        </Grid>
-
-        {assessmentStatus === 'complete' && (
-          <>
-            <Typography paragraph className={commonStyles.margintop1}>
-              Assessment completed. To continue please select from the following
-              option(s):
-            </Typography>
-            <Grid container direction="row" spacing={2}>
-              {workflowDefinitions.map(workflow => (
-                <Grid item>
-                  <Card
-                    key={workflow.name}
-                    raised
-                    className={styles.applicationCard}
-                  >
-                    <CardMedia>
-                      <ItemCardHeader
-                        title={workflow.name}
-                        classes={{ root: styles.applicationHeader }}
-                      />
-                    </CardMedia>
-                    <CardContent>{workflow.description}</CardContent>
-                    <CardActions>
-                      <Button
-                        id={workflow.id}
-                        variant="text"
-                        color="primary"
-                        href={`/parodos/onboarding/${project?.id}/${workflow.id}/new/`}
-                      >
-                        START
-                      </Button>
-                    </CardActions>
-                  </Card>
-                </Grid>
-              ))}
+                <Button
+                  type="submit"
+                  disabled={disableForm}
+                  variant="contained"
+                  color="primary"
+                >
+                  {inProgress ? 'IN PROGRESS' : 'START ASSESSMENT'}
+                </Button>
+              </Form>
             </Grid>
-          </>
-        )}
-      </Grid>
+            <Grid item xs={12}>
+              {assessmentStatus === 'complete' && project && (
+                <WorkflowDefinitions
+                  project={project}
+                  workflowDefinitions={workflowDefinitions.filter(
+                    workflowDefinition =>
+                      // TODO: is following correct?
+                      !['ASSESSMENT', 'CHECKER'].includes(
+                        workflowDefinition.type,
+                      ),
+                  )}
+                />
+              )}
+            </Grid>
+          </Grid>
+        </InfoCard>
+      )}
     </ParodosPage>
   );
-};
-
-export const Workflow: React.FC = props => (
-  <WorkflowParametersContextProvider>
-    <WorkflowImpl {...props} />
-  </WorkflowParametersContextProvider>
-);
+}
